requirements.txt
numpy==1.24.3
Pillow==9.5.0
opencv-python==4.7.0.72
--extra-index-url https://download.pytorch.org/whl/cu117
torch==2.0.1
--extra-index-url https://download.pytorch.org/whl/cu117
torchvision==0.15.2